# Docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: UMIG API V2
  version: 2.0.0
  description: |
    REST API for the Unified Migration and Implementation Governance (UMIG) application.

    This API follows RESTful principles and is implemented using ScriptRunner Groovy scripts. It supersedes the V1 API.
    Architectural principles are documented in ADR-017.
servers:
  - url: /rest/scriptrunner/latest/custom
    description: ScriptRunner Custom REST Endpoints (relative to Confluence base URL)
...
paths:
...
  /teams:
    get:
      tags:
        - Teams
      summary: Get teams with optional hierarchical filtering
      description: |
        Retrieves a list of teams. Can be filtered by hierarchical entities to show only teams 
        involved in specific migrations, iterations, plans, sequences, or phases.
      operationId: getAllTeams
      parameters:
        - name: migrationId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter teams by migration ID
        - name: iterationId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter teams by iteration ID
        - name: planId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter teams by plan instance ID
        - name: sequenceId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter teams by sequence instance ID
        - name: phaseId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter teams by phase instance ID
      responses:
        '200':
          description: A list of teams.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Teams
      summary: Create a new team
      description: Creates a new team with the provided details.
      operationId: createTeam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamCreateRequest'
      responses:
        '201':
          description: Team created successfully.
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /teams/{id}:
    get:
      tags:
        - Teams
      summary: Get a team by ID
      description: Retrieves a single team by its unique ID.
      operationId: getTeamById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The requested team.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Teams
      summary: Update a team
      description: Updates an existing team's details.
      operationId: updateTeam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamCreateRequest'
      responses:
        '204':
          description: Team updated successfully.
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Teams
      summary: Delete a team
      description: Deletes a team by its unique ID.
      operationId: deleteTeam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Team deleted successfully.
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /teams/{teamId}/users/{userId}:
    put:
      tags:
        - Teams
      summary: Add a user to a team
      description: Adds a user to a team, creating the membership link.
      operationId: addUserToTeam
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: User added to team successfully.
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Teams
      summary: Remove a user from a team
      description: Removes a user from a team, deleting the membership link.
      operationId: removeUserFromTeam
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: User removed from team successfully.
        '404':
          $ref: '#/components/responses/NotFound'
  /teams/{id}/applications:
    get:
      tags:
        - Teams
      summary: Get applications associated with a team
      description: Retrieves a list of applications that are associated with the specified team.
      operationId: getTeamApplications
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A list of applications associated with the team
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Application'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /teams/{teamId}/applications/{applicationId}:
    put:
      tags:
        - Teams
      summary: Add an application to a team
      description: Creates an association between a team and an application.
      operationId: addApplicationToTeam
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
        - name: applicationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Application added to team successfully.
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Application is already associated with this team.
    delete:
      tags:
        - Teams
      summary: Remove an application from a team
      description: Removes the association between a team and an application.
      operationId: removeApplicationFromTeam
      parameters:
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
        - name: applicationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Application removed from team successfully.
        '404':
          $ref: '#/components/responses/NotFound'
  /environments:
    get:
      tags:
        - Environments
      summary: Get all environments with counts
      description: |
        Retrieves a list of all environments with their associated application and iteration counts.
        Supports pagination, sorting, and search for admin GUI.
      operationId: getAllEnvironments
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
          description: Page number for pagination
        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
          description: Page size for pagination
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Search term to filter environments
        - name: sort
          in: query
          required: false
          schema:
            type: string
          description: Field to sort by
        - name: direction
          in: query
          required: false
          schema:
            type: string
            enum: [asc, desc]
          description: Sort direction
      responses:
        '200':
          description: A list of environments with counts
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/EnvironmentWithCounts'
                  - $ref: '#/components/schemas/PaginatedEnvironments'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Environments
      summary: Create a new environment
      description: Creates a new environment with the provided details.
      operationId: createEnvironment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvironmentCreateRequest'
      responses:
        '201':
          description: Environment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Environment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /environments/roles:
    get:
      tags:
        - Environments
      summary: Get all environment roles
      description: Retrieves a list of all available environment roles.
      operationId: getEnvironmentRoles
      responses:
        '200':
          description: A list of environment roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EnvironmentRole'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /environments/{id}/applications/{appId}:
    post:
      tags:
        - Environments
      summary: Associate an application with an environment
      description: Creates an association between an environment and an application.
      operationId: associateApplication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: appId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Application associated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Environments
      summary: Remove application association
      description: Removes the association between an environment and an application.
      operationId: disassociateApplication
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: appId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Application disassociated successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    Team:
      type: object
      properties:
        tms_id:
          type: integer
          description: Unique identifier for the team
        tms_name:
          type: string
          description: Name of the team
        tms_email:
          type: string
          format: email
          description: Email address for the team
        tms_description:
          type: string
          description: Description of the team
...
    TeamCreateRequest:
      type: object
      required:
        - tms_name
        - tms_email
      properties:
        tms_name:
          type: string
          description: Name of the team
        tms_email:
          type: string
          format: email
          description: Email address for the team
        tms_description:
          type: string
          description: Description of the team
...
    Application:
      type: object
      properties:
        app_id:
          type: integer
          description: Unique identifier for the application
        app_code:
          type: string
          description: Application code
        app_name:
          type: string
          description: Name of the application
        app_description:
          type: string
          description: Description of the application
...
  responses:
...
    NotFound:
      description: The specified resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: Invalid input, object invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: The request could not be completed due to a conflict with the current state of the resource (e.g., duplicate entry).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: An unexpected internal server error occurred.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'